pyproject.toml
[project]
name = "wordstat-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pretty-json-loguru>=0.2.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
